requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=5.0.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
celery==5.4.0
redis==5.0.8
beautifulsoup4==4.12.3
requests>=2.31.0
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0
